stock-pipeline/requirements.txt
yfinance>=0.2.40
pandas>=2.2.0
duckdb>=1.1.0
pyarrow>=15.0.0
pyyaml>=6.0
